Remove redundant, old, and duplicated color correction implementations
