--- app-backend/backsplash-core/src/main/scala/com/rasterfoundry/backsplash/color/ColorCorrect.scala
@@ -1,322 +1,0 @@
-package com.rasterfoundry.backsplash.color
-
-import io.circe.generic.JsonCodec
-import geotrellis.raster._
-import geotrellis.raster.histogram.Histogram
-import com.typesafe.scalalogging.LazyLogging
-import org.apache.commons.math3.util.FastMath
-import spire.syntax.cfor._
-
-/**
-  * Usage of Approximations.{pow | exp} functions can allow to speed up this function on 10 - 15ms.
-  * We can consider these functions usages in case of real performance issues caused by a long color correction.
-  */
-object ColorCorrect extends LazyLogging {
-  import functions.SaturationAdjust._
-  import functions.SigmoidalContrast._
-
-  final case class LayerClipping(redMin: Int,
-                                 redMax: Int,
-                                 greenMin: Int,
-                                 greenMax: Int,
-                                 blueMin: Int,
-                                 blueMax: Int)
-  sealed trait ClipValue
-  final case class ClipBounds(min: Int, max: Int) extends ClipValue
-  final case class MaybeClipBounds(maybeMin: Option[Int], maybeMax: Option[Int])
-      extends ClipValue
-  final case class ClippingParams(band: Int, bounds: ClipValue)
-
-  @JsonCodec
-  final case class Params(redBand: Int,
-                          greenBand: Int,
-                          blueBand: Int,
-                          gamma: BandGamma,
-                          bandClipping: PerBandClipping,
-                          tileClipping: MultiBandClipping,
-                          sigmoidalContrast: SigmoidalContrast,
-                          saturation: Saturation) {
-    def getGamma: Map[Int, Option[Double]] = Map(
-      0 -> (if (gamma.enabled) gamma.redGamma else None),
-      1 -> (if (gamma.enabled) gamma.greenGamma else None),
-      2 -> (if (gamma.enabled) gamma.blueGamma else None)
-    )
-
-    def colorCorrect(tile: MultibandTile,
-                     hist: Seq[Histogram[Double]],
-                     nodataValue: Option[Double]): MultibandTile = {
-      val indexedHist = hist.toIndexedSeq
-      val rgbHist = Seq(redBand, greenBand, blueBand) map { indexedHist(_) }
-      ColorCorrect(tile, rgbHist, this, nodataValue)
-    }
-
-    def withBands(red: Int, green: Int, blue: Int) =
-      this.copy(redBand = red, greenBand = green, blueBand = blue)
-  }
-
-  @inline def normalizeAndClampAndGammaCorrectPerPixel(
-      z: Double,
-      oldMin: Int,
-      oldMax: Int,
-      newMin: Int,
-      newMax: Int,
-      gammaOpt: Option[Double],
-      noDataValue: Option[Double]
-  ): Int = {
-    if (isData(z)) {
-      val dNew = newMax - newMin
-      val dOld = oldMax - oldMin
-
-      if (noDataValue.map(_ == z).getOrElse(false)) {
-        0
-      } else if (dOld == 0) {
-        // When dOld is nothing (normalization is meaningless in this context), we still need to clamp
-        val v = {
-          if (z > newMax) newMax
-          else if (z < newMin) newMin
-          else z.toInt
-        }
-
-        gammaOpt match {
-          case None => v
-          case Some(gamma) =>
-            clampColor {
-              val gammaCorrection = 1 / gamma
-              (255 * FastMath.pow(v / 255.0, gammaCorrection)).toInt
-            }
-        }
-      } else {
-        val v = {
-          val scaled = (((z - oldMin) * dNew) / dOld) + newMin
-
-          if (scaled > newMax) newMax
-          else if (scaled < newMin) newMin
-          else scaled.toInt
-        }
-
-        gammaOpt match {
-          case None => v
-          case Some(gamma) =>
-            clampColor {
-              val gammaCorrection = 1 / gamma
-              (255 * FastMath.pow(v / 255.0, gammaCorrection)).toInt
-            }
-        }
-      }
-    } else z.toInt
-  }
-
-  val rgbBand: (Option[Int], Option[Int], Int) => Some[Int] =
-    (specificBand: Option[Int], allBands: Option[Int], tileDefault: Int) =>
-      specificBand.fold(allBands)(Some(_)).fold(Some(tileDefault))(x => Some(x))
-
-  def complexColorCorrect(rgbTile: MultibandTile, saturation: Saturation)(
-      layerNormalizeArgs: Map[Int, ClipBounds],
-      gammas: Map[Int, Option[Double]]
-  )(sigmoidalContrast: SigmoidalContrast)(
-      colorCorrectArgs: Map[Int, MaybeClipBounds],
-      tileClipping: MultiBandClipping,
-      noDataValue: Option[Double]
-  ): MultibandTile = {
-    val (red, green, blue) = (rgbTile.band(0), rgbTile.band(1), rgbTile.band(2))
-    val (gr, gg, gb) = (gammas(0), gammas(1), gammas(2))
-    val (nred, ngreen, nblue) = (
-      ArrayTile.alloc(rgbTile.cellType, rgbTile.cols, rgbTile.rows),
-      ArrayTile.alloc(rgbTile.cellType, rgbTile.cols, rgbTile.rows),
-      ArrayTile.alloc(rgbTile.cellType, rgbTile.cols, rgbTile.rows)
-    )
-
-    val ClipBounds(rmin, rmax) = layerNormalizeArgs(0)
-    val ClipBounds(gmin, gmax) = layerNormalizeArgs(1)
-    val ClipBounds(bmin, bmax) = layerNormalizeArgs(2)
-
-    val tileMin = 1
-    val (rclipMin, rclipMax, rnewMin, rnewMax) = (rmin, rmax, tileMin, 255)
-    val (gclipMin, gclipMax, gnewMin, gnewMax) = (gmin, gmax, tileMin, 255)
-    val (bclipMin, bclipMax, bnewMin, bnewMax) = (bmin, bmax, tileMin, 255)
-
-    val sigmoidal: Double => Double =
-      (
-        sigmoidalContrast.enabled,
-        sigmoidalContrast.alpha,
-        sigmoidalContrast.beta
-      ) match {
-        case (true, Some(a), Some(b)) => localTransform(rgbTile.cellType, a, b)
-        case _                        => identity
-      }
-
-    val (clipr, clipg, clipb): (Int => Int, Int => Int, Int => Int) = {
-      val MaybeClipBounds(mrmin, mrmax) = colorCorrectArgs(0)
-      val MaybeClipBounds(mgmin, mgmax) = colorCorrectArgs(1)
-      val MaybeClipBounds(mbmin, mbmax) = colorCorrectArgs(2)
-
-      val (mrclipMin, mrclipMax) = (
-        rgbBand(mrmin, tileClipping.min, 0).get,
-        rgbBand(mrmax, tileClipping.max, 255).get
-      )
-      val (mgclipMin, mgclipMax) = (
-        rgbBand(mgmin, tileClipping.min, 0).get,
-        rgbBand(mgmax, tileClipping.max, 255).get
-      )
-      val (mbclipMin, mbclipMax) = (
-        rgbBand(mbmin, tileClipping.min, 0).get,
-        rgbBand(mbmax, tileClipping.max, 255).get
-      )
-
-      @inline def clipBands(z: Int, min: Int, max: Int): Int = {
-        if (isData(z) && z > max) 255
-        else if (isData(z) && z < min) 0
-        else z
-      }
-
-      (
-        clipBands(_, mrclipMin, mrclipMax),
-        clipBands(_, mgclipMin, mgclipMax),
-        clipBands(_, mbclipMin, mbclipMax)
-      )
-    }
-
-    logger.trace(
-      s"Red (Clip Min: ${rclipMin}, Max: ${rclipMax}) (New Min: ${rnewMin}, ${rnewMax})")
-
-    /** In this case for some reason with this func wrap it works faster ¯\_(ツ)_/¯ (it was micro benchmarked) */
-    lazyWrapper {
-      cfor(0)(_ < rgbTile.cols, _ + 1) { col =>
-        cfor(0)(_ < rgbTile.rows, _ + 1) { row =>
-          val (r, g, b) =
-            (
-              ColorCorrect.normalizeAndClampAndGammaCorrectPerPixel(
-                red.getDouble(col, row),
-                rclipMin,
-                rclipMax,
-                rnewMin,
-                rnewMax,
-                gr,
-                noDataValue
-              ),
-              ColorCorrect.normalizeAndClampAndGammaCorrectPerPixel(
-                green.getDouble(col, row),
-                gclipMin,
-                gclipMax,
-                gnewMin,
-                gnewMax,
-                gg,
-                noDataValue
-              ),
-              ColorCorrect.normalizeAndClampAndGammaCorrectPerPixel(
-                blue.getDouble(col, row),
-                bclipMin,
-                bclipMax,
-                bnewMin,
-                bnewMax,
-                gb,
-                noDataValue
-              )
-            )
-
-          val (nr, ng, nb) = (saturation.enabled, saturation.saturation) match {
-            case (true, Some(cf)) =>
-              val (hue, chroma, luma) = rgbToHcluma(r, g, b)
-              val newChroma = scaleChroma(chroma, cf)
-              val (nr, ng, nb) = hclumaToRgb(hue, newChroma, luma)
-              (nr, ng, nb)
-
-            case _ => (r, g, b)
-          }
-
-          nred.set(col, row, clipr(sigmoidal(nr).toInt))
-          ngreen.set(col, row, clipg(sigmoidal(ng).toInt))
-          nblue.set(col, row, clipb(sigmoidal(nb).toInt))
-        }
-      }
-    }
-
-    MultibandTile(nred, ngreen, nblue)
-  }
-
-  def apply(rgbTile: MultibandTile,
-            rgbHist: Seq[Histogram[Double]],
-            params: Params,
-            noDataValue: Option[Double]): MultibandTile = {
-    val _rgbTile = rgbTile
-    val _rgbHist = rgbHist
-    val gammas = params.getGamma
-
-    val layerRgbClipping = {
-      val range = 1 until 255
-      var isCorrected = true
-      val iMaxMin: Array[(Int, Int)] = Array.ofDim(3)
-      cfor(0)(_ < _rgbHist.length, _ + 1) { index =>
-        val hst = _rgbHist(index)
-        val imin = hst.minValue().map(_.toInt).getOrElse(0)
-        val imax = hst.maxValue().map(_.toInt).getOrElse(255)
-        logger.trace(s"Histogram Min/Max: ${imin}/${imax}")
-        iMaxMin(index) = (imin, imax)
-        isCorrected = {
-          if (range.contains(imin) && range.contains(imax)) true
-          else false
-        }
-      }
-
-      if (!isCorrected) {
-        val (rmin, rmax) = iMaxMin(0)
-        val (gmin, gmax) = iMaxMin(1)
-        val (bmin, bmax) = iMaxMin(2)
-        LayerClipping(rmin, rmax, gmin, gmax, bmin, bmax)
-      } else LayerClipping(0, 255, 0, 255, 0, 255)
-    }
-
-    val layerNormalizeArgs: Map[Int, ClipBounds] = Map(
-      0 -> ClipBounds(layerRgbClipping.redMin, layerRgbClipping.redMax),
-      1 -> ClipBounds(layerRgbClipping.greenMin, layerRgbClipping.greenMax),
-      2 -> ClipBounds(layerRgbClipping.blueMin, layerRgbClipping.blueMax)
-    )
-
-    val colorCorrectArgs: Map[Int, MaybeClipBounds] = Map(
-      0 -> MaybeClipBounds(
-        params.bandClipping.redMin,
-        params.bandClipping.redMax
-      ),
-      1 -> MaybeClipBounds(
-        params.bandClipping.greenMin,
-        params.bandClipping.greenMax
-      ),
-      2 -> MaybeClipBounds(
-        params.bandClipping.blueMin,
-        params.bandClipping.blueMax
-      )
-    )
-
-    logger.trace(s"ColorCorrectArgs: ${colorCorrectArgs}")
-    logger.trace(s"Layer Normalize Args: ${layerNormalizeArgs}")
-    complexColorCorrect(_rgbTile, params.saturation)(
-      layerNormalizeArgs,
-      gammas
-    )(params.sigmoidalContrast)(colorCorrectArgs,
-                                params.tileClipping,
-                                noDataValue)
-  }
-
-  @inline def clampColor(z: Int): Int = {
-    if (z < 0) 0
-    else if (z > 255) 255
-    else z
-  }
-
-  private def lazyWrapper[T](f: => T): T = f
-
-  def paramsFromBandSpecOnly(
-      redBand: Int,
-      greenBand: Int,
-      blueBand: Int
-  ): Params = Params(
-    redBand,
-    greenBand,
-    blueBand,
-    BandGamma(false, None, None, None),
-    PerBandClipping(false, None, None, None, None, None, None),
-    MultiBandClipping(false, None, None),
-    SigmoidalContrast(false, None, None),
-    Saturation(false, None)
-  )
-}
