+++ app-backend/backsplash-core/src/main/scala/com/rasterfoundry/backsplash/color/ColorRampMosaic.scala
@@ -1,7 +1,7 @@
 package com.rasterfoundry.backsplash.color
 
+import com.rasterfoundry.datamodel.SingleBandOptions
 import com.rasterfoundry.backsplash.error._
-
 import com.typesafe.scalalogging.LazyLogging
 import geotrellis.raster._
 import geotrellis.raster.histogram._
