+++ app-backend/backsplash-core/src/main/scala/com/rasterfoundry/backsplash/color/OgcStyles.scala
@@ -1,6 +1,7 @@
 package com.rasterfoundry.backsplash.color
 
-import com.rasterfoundry.datamodel.ColorComposite
+import com.rasterfoundry.datamodel.{ColorComposite, SingleBandOptions}
+import com.rasterfoundry.common.color.ColorCorrect
 import geotrellis.raster._
 import geotrellis.raster.histogram._
 import geotrellis.server.ogc.OutputFormat.Png
